refactor: extract RM_CSHARP to rm-csharp.js
